+++ src/std/gui/src/lib.rs
@@ -1,4 +1,5 @@
 use nstd_events::{NSTDEventLoop, NSTDWindowID};
+use nstd_image::NSTDImage;
 use std::{
     ffi::CStr,
     os::raw::{c_char, c_double, c_int},
@@ -265,22 +266,19 @@ pub unsafe extern "C" fn nstd_std_gui_window_is_maximized(window: NSTDWindow) ->
 /// Sets a window's icon image.
 /// Parameters:
 ///     `NSTDWindow window` - The window.
-///     `const NSTDByte *const raw` - Raw image data, must be in RGBA format or null.
-///     `const NSTDUInt32 width` - The image width.
-///     `const NSTDUInt32 height` - The image height.
+///     `const NSTDImage *const img` - The icon image, null for default.
 /// Returns: `int errc` - Nonzero on error.
 #[cfg_attr(feature = "clib", no_mangle)]
 pub unsafe extern "C" fn nstd_std_gui_window_set_icon(
     window: NSTDWindow,
-    raw: *const u8,
-    width: u32,
-    height: u32,
+    img: *const NSTDImage,
 ) -> c_int {
-    let icon = match !raw.is_null() {
+    let icon = match !img.is_null() {
         true => {
             const RGBA_COMPONENTS: u32 = 4;
-            let raw = slice::from_raw_parts(raw, (RGBA_COMPONENTS * width * height) as usize);
-            match Icon::from_rgba(raw.to_vec(), width, height) {
+            let size = (RGBA_COMPONENTS * (*img).width * (*img).height) as usize;
+            let raw = slice::from_raw_parts((*img).raw, size);
+            match Icon::from_rgba(raw.to_vec(), (*img).width, (*img).height) {
                 Ok(icon) => Some(icon),
                 _ => return 1,
             }