+++ src/std/gui/CHANGELOG.md
@@ -1,4 +1,6 @@
 # nstd_gui
+## Latest
+- `nstd_std_gui_window_set_icon` now takes a pointer to an `NSTDImage`.
 ## 0.2.0
 - `clib` feature.
 ## 0.1.0